Refactor: Changed I2NP message structure - header is not really suitable to be super class
